add : accordion, productId, showNewProductForm state hooks 구현
