+++ src/refactoring/features/admin/hooks/useNewProductFormToggle.ts
@@ -1,0 +1,10 @@
+import { useState } from 'react';
+
+export const useNewProductFormToggle = () => {
+  const [showNewProductForm, setShowNewProductForm] = useState(false);
+
+  const openNewProductForm = () => setShowNewProductForm(true);
+  const closeNewProductForm = () => setShowNewProductForm(false);
+  const toggleNewProductForm = () => setShowNewProductForm((prev) => !prev);
+  return { showNewProductForm, setShowNewProductForm, openNewProductForm, closeNewProductForm, toggleNewProductForm };
+};
